fix a few issues with tickers_value

+++ table/get_context.cc
@@ -91,7 +91,7 @@ void GetContext::RecordCounters(Tickers ticker, size_t val) {
   if (ticker == Tickers::TICKER_ENUM_MAX) {
     return;
   }
-  tickers_value[ticker] += static_cast<uint32_t>(val);
+  tickers_value[ticker] += static_cast<uint64_t>(val);
 }
 
 bool GetContext::SaveValue(const ParsedInternalKey& parsed_key,
